update modal ui and split stake modal in subcomponents
